feat: Add file creation logic to compliance check code

+++ src/compliance-check/main.ts
@@ -2,22 +2,25 @@ import tl = require("azure-pipelines-task-lib/task");
 
 async function run() {
     try {
-
-        const inputString: string | undefined = tl.getInput("filename", true);
-        console.log("Reading immediate values...." + inputString);
-        console.log("finished execution", inputString);
-
         const fs = require("fs");
         const path = require("path");
-        const fileName = inputString;
-        const filePath = path.join(__dirname, fileName);
-        const currentTime = new Date().toISOString();
-        fs.writeFileSync(filePath, currentTime);
-        console.log("File created successfully");
+
+        const fileName: string | undefined = tl.getInput("filename", true);
+        const filePath: string | undefined = tl.getInput("filepath", true);
+
+        console.log("File name provided by user...." + fileName);
+        console.log("File path provided by user...." + filePath);
+        // const filePath = path.join(__dirname, fileName);
+
+        if (!fs.existsSync(filePath)) {
+            const currentTime = new Date().toISOString();
+            fs.writeFileSync(filePath, currentTime);
+            console.log("File was not found, created a new file with current time....");
+        }
 
         // now upload this file as attachment
         const attachmentType = "Distributedtask.Core.Summary";
-        const attachmentName = "ComplianceCheck";
+        const attachmentName = fileName;
         const attachmentLocation = filePath;
         console.log("Uploading attachment....");
         tl.command("task.addattachment", { type: attachmentType, name: attachmentName }, attachmentLocation);
